Move spring-websocket into spring-boot generator
