+++ generators/spring-boot/generators/websocket/cleanup.ts
