+++ generators/spring-boot/generators/websocket/templates/src/main/java/_package_/config/WebsocketConfiguration.java.ejs
