+++ generators/spring-boot/generators/websocket/templates/src/main/java/_package_/config/WebsocketSecurityConfiguration.java.ejs
