+++ generators/spring-boot/generators/websocket/templates/src/main/java/_package_/web/websocket/dto/ActivityDTO.java.ejs
