--- generators/spring-websocket/generator.spec.ts
@@ -1,28 +1,0 @@
-import { before, describe, expect, it } from 'esmocha';
-import { basename, dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-import { defaultHelpers as helpers, result } from '../../lib/testing/index.ts';
-import { shouldSupportFeatures, testBlueprintSupport } from '../../test/support/tests.js';
-
-import Generator from './index.ts';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-const generator = basename(__dirname);
-
-describe(`generator - ${generator}`, () => {
-  shouldSupportFeatures(Generator);
-  describe('blueprint support', () => testBlueprintSupport(generator));
-
-  describe('with default config', () => {
-    before(async () => {
-      await helpers.runJHipster(generator).withJHipsterConfig();
-    });
-
-    it('should match files snapshot', () => {
-      expect(result.getStateSnapshot()).toMatchSnapshot();
-    });
-  });
-});
