+++ lib/index.spec.ts
@@ -111,6 +111,7 @@ describe('getGeneratorNamespaces', () => {
   "spring-boot:bootstrap",
   "spring-boot:jwt",
   "spring-boot:oauth2",
+  "spring-boot:websocket",
   "spring-cache",
   "spring-cloud-stream",
   "spring-cloud-stream:kafka",
@@ -122,7 +123,6 @@ describe('getGeneratorNamespaces', () => {
   "spring-data-mongodb",
   "spring-data-neo4j",
   "spring-data-relational",
-  "spring-websocket",
   "upgrade",
   "vue",
   "vue:bootstrap",
